+++ packages/atomic/src/components/facets/base-facet/base-facet.tsx
@@ -55,7 +55,7 @@ export const BaseFacet: FunctionalComponent<BaseFacetProps> = (
   ) : null;
 
   return (
-    <div class="facet" part="facet">
+    <div class="facet mb-4" part="facet">
       <button
         class={
           'facet-button border-solid bg-transparent px-4 h-9 outline-none focus:outline-none lg:hidden cursor-pointer ' +
@@ -69,7 +69,7 @@ export const BaseFacet: FunctionalComponent<BaseFacetProps> = (
       </button>
       <div
         class={
-          'content box-border lg:block h-screen w-screen lg:h-auto lg:w-auto fixed object-left-top bg-white top-0 left-0 lg:static p-3 z-10 ' +
+          'content box-border p-3 lg:p-0 lg:block h-screen w-screen lg:h-auto lg:w-auto fixed object-left-top bg-background top-0 left-0 lg:static z-10 ' +
           (props.controller.state.isExpanded ? 'block' : 'hidden')
         }
       >
@@ -77,7 +77,7 @@ export const BaseFacet: FunctionalComponent<BaseFacetProps> = (
           <span class="font-semibold text-on-background text-base lg:text-sm">
             {props.label}
           </span>
-          <span class="flex flex-row block ml-auto">
+          <span class="flex flex-row ml-auto">
             {resetButton}
             {closeButton}
           </span>