+++ packages/atomic/src/components/result-template-components/atomic-field-condition/atomic-field-condition.tsx
@@ -1,13 +1,10 @@
-import {Component, Prop, h, Method, Element} from '@stencil/core';
+import {Component, Prop, h, Element} from '@stencil/core';
 import {
   Result,
   ResultTemplateCondition,
   ResultTemplatesHelpers,
 } from '@coveo/headless';
-import {
-  ResultContext,
-  ResultContextRenderer,
-} from '../result-template-decorators';
+import {ResultContext} from '../result-template-decorators';
 import {MapProp} from '../../../utils/props-utils';
 
 @Component({
@@ -23,7 +20,6 @@ export class AtomicFieldCondition {
   @MapProp() mustMatch: Record<string, string[]> = {};
   @MapProp() mustNotMatch: Record<string, string[]> = {};
 
-  private fields: string[] = [];
   private shouldBeRemoved = false;
 
   @ResultContext() private result!: Result;
@@ -31,7 +27,6 @@ export class AtomicFieldCondition {
   public componentWillLoad() {
     if (this.ifDefined) {
       const fieldNames = this.ifDefined.split(',');
-      this.fields.push(...fieldNames);
       this.conditions.push(
         ResultTemplatesHelpers.fieldsMustBeDefined(fieldNames)
       );
@@ -39,7 +34,6 @@ export class AtomicFieldCondition {
 
     if (this.ifNotDefined) {
       const fieldNames = this.ifNotDefined.split(',');
-      this.fields.push(...fieldNames);
       this.conditions.push(
         ResultTemplatesHelpers.fieldsMustNotBeDefined(fieldNames)
       );
@@ -49,7 +43,6 @@ export class AtomicFieldCondition {
       this.conditions.push(
         ResultTemplatesHelpers.fieldMustMatch(field, this.mustMatch[field])
       );
-      this.fields.push(field);
     }
 
     for (const field in this.mustNotMatch) {
@@ -59,11 +52,9 @@ export class AtomicFieldCondition {
           this.mustNotMatch[field]
         )
       );
-      this.fields.push(field);
     }
   }
 
-  @ResultContextRenderer
   public render() {
     if (!this.conditions.every((condition) => condition(this.result))) {
       this.shouldBeRemoved = true;
@@ -75,10 +66,5 @@ export class AtomicFieldCondition {
 
   public componentDidLoad() {
     this.shouldBeRemoved && this.host.remove();
-  }
-
-  @Method()
-  public async getFields() {
-    return this.fields;
   }
 }