+++ packages/atomic/stencil.config.ts
@@ -17,7 +17,6 @@ const isDevWatch: boolean =
 export const config: Config = {
   namespace: 'atomic',
   taskQueue: 'async',
-  globalStyle: 'src/globals/global.pcss',
   outputTargets: [
     {
       type: 'dist',