+++ packages/headless/src/controllers/index.ts
@@ -9,7 +9,13 @@ export {
   buildRelevanceInspector,
 } from './relevance-inspector/headless-relevance-inspector';
 
-export {Context, ContextState, buildContext} from './context/headless-context';
+export {
+  Context,
+  ContextState,
+  ContextValue,
+  ContextPayload,
+  buildContext,
+} from './context/headless-context';
 
 export {
   DidYouMean,
@@ -73,8 +79,8 @@ export {
 } from './pager/headless-pager';
 
 export {
-  QueryErrorState,
   QueryError,
+  QueryErrorState,
   buildQueryError,
 } from './query-error/headless-query-error';
 
@@ -225,3 +231,5 @@ export {
   SearchStatusState,
   buildSearchStatus,
 } from './search-status/headless-search-status';
+
+export {ErrorPayload} from './controller/error-payload';